restore automatic choosing of trunclenf & trunclenr
