+++ conf/modules.config
@@ -28,11 +28,16 @@ params {
         'cutadapt' {
             args          = ""
         }
+        'dada_quality' {
+            args          = "5e+06"
+            publish_files = ['pdf':'']
+        }
         'dada_filter_and_trim' {
             args          = "maxN = 0"
         }
         'dada_error_model' {
             args          = "randomize=FALSE"
+            publish_files = ['pdf':'']
         }
         'dada_denoising' {
             args          = ""
