+++ conf/test_manifest.config
@@ -23,7 +23,5 @@ params {
   outdir = "./results"
   temp_dir = "./results/tmp_dir"
   input = "https://github.com/nf-core/test-datasets/raw/ampliseq/testdata/Manifest.tsv"
-  trunclenf = 0
-  trunclenr = 0
   dada_ref_taxonomy = "https://zenodo.org/record/4310151/files/rdp_train_set_18.fa.gz"
 }
